+++ src/main.rs
@@ -668,6 +668,10 @@ fn dispatch_common_subcommand(
             let result = commands::list_federation_upstreams(client);
             res_handler.tabular_result(result)
         }
+        ("federation", "disable_tls_peer_verification_for_all_upstreams") => {
+            let result = commands::disable_tls_peer_verification_for_all_federation_upstreams(client);
+            res_handler.no_output_on_success(result);
+        }
         ("get", "messages") => {
             let result = commands::get_messages(client, &vhost, second_level_args);
             res_handler.tabular_result(result)